backend/requirements.txt
# 수정일: 2026-01-20
# 수정내용: Django 백엔드 핵심 라이브러리 정의

django
django-environ
psycopg2-binary
djangorestframework
django-cors-headers
drf-spectacular
